Move overlapping logic to useFieldValue hook
