+++ packages/dataviews/src/dataforms-layouts/regular/index.tsx
@@ -16,7 +16,7 @@ import type { Form, FieldLayoutProps } from '../../types';
 import DataFormContext from '../../components/dataform-context';
 import { DataFormLayout } from '../data-form-layout';
 import { isCombinedField } from '../is-combined-field';
-import { MIXED_VALUE } from '../../constants';
+import useFieldValue from '../use-field-value';
 
 function Header( { title }: { title: string } ) {
 	return (
@@ -60,27 +60,7 @@ export default function FormRegularField< Item >( {
 		};
 	}, [ field ] );
 
-	const fieldValue = useMemo( () => {
-		const fieldDefinition = fields.find(
-			( fieldDef ) => fieldDef.id === field.id
-		);
-		if ( ! fieldDefinition ) {
-			return undefined;
-		}
-		if ( Array.isArray( data ) ) {
-			const [ firstRecord, ...remainingRecords ] = data;
-			const firstValue = fieldDefinition.getValue( {
-				item: firstRecord,
-			} );
-			const intersects = remainingRecords.every( ( item ) => {
-				return fieldDefinition.getValue( { item } ) === firstValue;
-			} );
-			return intersects ? firstValue : MIXED_VALUE;
-		}
-		return fieldDefinition.getValue( {
-			item: data,
-		} );
-	}, [ data, fields, field.id ] );
+	const fieldValue = useFieldValue( data, field.id );
 
 	if ( isCombinedField( field ) ) {
 		return (
